requirements.txt
Pillow==10.1.0
reportlab==4.0.7
PyPDF2==3.0.1
tkcalendar==1.6.1
openpyxl==3.11.0
